services/alert-service/requirements.txt
kafka-python>=2.0.2
redis>=5.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
